requirements.txt
streamlit
numpy
pillow
tensorflow>=2.8.0
torch
torchvision
Pillow
